Create types directory to store all typescript types
