+++ ClientApp/services/TodoAppDataService.ts
@@ -1,31 +1,6 @@
 import http from '../src/http-common';
+import type { CreateTodo, DeleteTodo, Todo, UpdateTodo } from '../src/types/todo-types';
 
-export interface Todo {
-	id: number;
-	title: string;
-	isCompleted: boolean;
-	userId: number;
-	lastUpdate: Date;
-	lastStatusUpdate: Date;
-}
-
-export interface CreateTodo {
-	title: string;
-	isCompleted: boolean;
-	userId: number;
-}
-
-export interface UpdateTodo {
-	id: number;
-	title: string;
-	isCompleted: boolean;
-	userId: number;
-}
-
-export interface DeleteTodo {
-	id: number;
-	userId: number;
-}
 class TodoAppDataService {
 	baseUrl = '/TodoTask';
 	async getAll(userId: number) {
