+++ ClientApp/src/types/auth-types.ts
@@ -1,0 +1,10 @@
+export interface Credentials {
+	username: string;
+	password: string;
+}
+
+export interface User {
+	id: number | string;
+	username: string;
+	token: string;
+}
